Fixed bug with trash stack. Added counters.
At the end of an interation, after displaying of Answers, a function is called to clear out the answer and trash stack. Preventign data from one run to persist in the next batch.

Also added counters to the displayed answers and trash. Totalling up the solutions found and dead branches sent to trash.